tests/submission/W3C/domhtml/HTMLInputElement14.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - INPUT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-input-element.html#dom-input-src" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<table border="BORDER" cellpadding="15" summary="Table 1">
<tbody><tr align="center">
<td valign="top">Under a FORM control
<form action="./files/getData.pl" id="form1" method="post">
<table border="15" summary="Table 2">
<tbody><tr>
<td>
<label accesskey="b" for="input1">Enter Your Password:</label>
</td>
<td>
<input alt="Password entry" dir="LTR" id="input1" maxlength="5" name="Password" readonly="READONLY" size="25" tabindex="8" type="password" value="Password">
</td>
</tr>
<tr>
<td>
<input accesskey="c" name="Radio1" type="RADIO" value="ReHire">ReHire
</td>
</tr>
<tr>
<td>
<input checked="CHECKED" name="Radio2" tabindex="9" type="RADIO" value="NewHire">NewHire
</td>
</tr>
<tr>
<td>Hours available to work</td>
<td>
<input align="bottom" checked="CHECKED" name="Check1" tabindex="10" type="CHECKBOX" value="EarlyMornings">EarlyMornings
<br>
<input id="input5" name="Check2" onclick="newId(this)" tabindex="11" type="CHECKBOX" value="AfterNoon">Afternoon
<br>
<input name="Check3" tabindex="12" type="CHECKBOX" value="Evenings">Evenings
<br>
<input disabled="DISABLED" name="Check4" tabindex="13" type="CHECKBOX" value="Closing">Closing
<br>
</td>
</tr>
<tr>
<td colspan="2">
<input name="SubmitImage" src="/support/images/yellow-rectangle.png" tabindex="14" type="IMAGE" usemap="#submit-map">
</td>
</tr>
<tr>
<td colspan="2">
<input accept="GIF,JPEG" name="FileControl" onselect="newTitle(this)" tabindex="15" title="old_title" type="FILE">
</td>
</tr>
</tbody></table>
</form>
</td>
</tr>
</tbody></table>


<script>
var domtest = async_test("The src attribute specifies the location of the image to decorate the     graphical submit button when the type has the value image.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vsrc;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("input");
      assert_equals(nodeList.length,9);
testNode = nodeList.item(7);
      vsrc = testNode.src;

      assert_uri_equals(document.location.protocol,null,document.location.host,"yellow-rectangle.png",null,null,null,true,vsrc);


    domtest.done();
}));

</script>
<p>The src attribute specifies the location of the image to decorate the
    graphical submit button when the type has the value image.
The src IDL attribute must reflect the respective content attribute of the same name.
    Retrieve the src attribute of the 8th INPUT element and examine
    its value.</p>
<div id="log"></div>
</body></html>
